✨💾 feat(komga): Add Komga configuration (#1436)
* ✨💾 feat(komga): Add Komga configuration

Adds the configuration files for the Komga media server application, including the
config.json and docker-compose.yml files. Komga is a media server for comics,
mangas, BDs, magazines, and eBooks, providing a convenient way to manage and
access this type of content.

* ✨ feat(komga): add installation tips to docker-compose

The changes in this commit add installation tips to the docker-compose file
for the Komga application. The tips provide users with a link to important
documentation that they should read before installing the application.
This change aims to improve the user experience and ensure that users
have the necessary information to successfully install and set up the
Komga application.

* ✨ feat(komga): upgrade Komga to latest version

Update Komga container image from 1.1.0 to 1.13.0 to
take advantage of the latest features and security
improvements in the Komga service.