✨ feat(Scrutiny): Enhance port mapping and device access (#1470)
* ✨ feat(Scrutiny): Enhance port mapping and device access

Changes the port mapping for the Scrutiny service to use different host ports (38080 and 38086) to
avoid potential conflicts with other services. Additionally, the changes map specific host devices
(e.g., /dev/sda and /dev/nvme0) to the container, allowing Scrutiny to access raw I/O devices.
This is necessary for Scrutiny to function properly. Finally, a healthcheck is added to ensure the
Scrutiny container is running correctly.

* 🔥 feat(docker-compose): use environment variable for data paths

Modify the docker-compose.yml file to use an environment variable for the
data paths instead of hardcoded values. This allows for more flexibility and
easier configuration of the Scrutiny application.