🐻 feat(note-mark-aio): Add new application configuration (#1495)
* 🐻 feat(note-mark-aio): Add new application configuration

Adds a new application configuration for the "note-mark-aio" application.
The changes include:
- Adding a `config.json` file with application metadata
- Adding a `docker-compose.yml` file with the service definition
- Adding a logo image for the application
The changes are made to provide a complete configuration for the
"note-mark-aio" application to be included in the Big Bear CasaOS
application catalog.

* ✨ docs(note-mark-aio): update installation instructions

Update the installation instructions for the Note Mark AIO
app in the docker-compose.yml file to point to the correct
community thread.