add-big-bear-flowise (#1574)
* ✨ feat(FloWise): Update Docker image and config

The changes update the Docker image and configuration for the FloWise
application. The Docker image has been changed from `flowise/flowise` to
`flowiseai/flowise`, which reflects the new image source. Additionally,
the `config.json` file has been updated to use the new image name.

* 📝 docs(docker-compose): update flowise installation instructions

Update the installation instructions for the Flowise app in the
docker-compose.yml file to reflect the correct community thread URL.
This ensures users have access to the latest information when
installing the Flowise app.