✨ feat(BigBearCasaOS): add RetroArch web player (#1622)
* ✨ feat(BigBearCasaOS): add RetroArch web player

This commit adds the RetroArch web player to the BigBearCasaOS applications.
The key changes are:

- Added a `docker-compose.yml` file to define the retroarch-web service
- Added a `config.json` file with metadata about the retroarch-web app
- Added a `logo.png` file for the retroarch-web app

These changes will allow users to easily deploy and use the RetroArch web player
on their BigBearCasaOS instances.

* 📝️ doc(retroarch-web): add tips section and refine ROOT_WWW_PATH env

- Adds a new `tips` section to the `docker-compose.yml` file with instructions for
  the user before installation
- Refines the `ROOT_WWW_PATH` environment variable to have a more descriptive
  container name and description

* ✨ feat(retroarch-web): update docker-compose instructions

Update the `before_install` instructions in the `retroarch-web`
docker-compose configuration to include a link to the BigBearTechWorld
community post about the added RetroArch Web Player.