✨ feat(bookstack): Update MariaDB data volume mount path (#1962)
* ✨ feat(bookstack): Update MariaDB data volume mount path

The changes in this commit update the MariaDB data volume mount path from `/var/lib/mysql` to `/config` in the BookStack Docker Compose configuration. This change aligns the volume mount path with the recommended best practices for container-based applications, where the `/config` directory is typically used to store persistent application data.

Additionally, the commit adds a new environment variable `APP_KEY` to the BookStack service, which is required for the application to function correctly. This environment variable is set to a base64-encoded value, which should be replaced with a unique value for the specific deployment.

* ✨ feat(bookstack): update docker-compose configuration

Update the BookStack docker-compose configuration to remove the
unnecessary environment variable and add a tips section with a link
to the documentation.

The changes made are:

- Removed the unnecessary `docker run` command from the environment
  variables section.
- Added a `tips` section with a link to the documentation for
  installing BookStack on BigBearCasaOS.