✨ feat(umbrel-os): Add tips section to docker-compose.yml
Adds a new "tips" section to the docker-compose.yml file for the
umbrel-os application. This includes a "before_install" tip that
directs users to read an important community post before
installing the application.

Also updates the name of the application from "umbrel" to
"big-bear-umbrel" to better reflect the customized nature of
this deployment.