umbrel-os-docker-compose-update (#2163)
* ✨ feat(umbrel-os): Add Umbrel OS configuration

Adds the necessary configuration files for the Umbrel OS application to be
deployed on the BigBear CasaOS platform. This includes the
`docker-compose.yml` file that defines the Umbrel service and its
configuration, as well as the `config.json` file that provides metadata
about the application.

* ✨ feat(umbrel-os): Add tips section to docker-compose.yml

Adds a new "tips" section to the docker-compose.yml file for the
umbrel-os application. This includes a "before_install" tip that
directs users to read an important community post before
installing the application.

Also updates the name of the application from "umbrel" to
"big-bear-umbrel" to better reflect the customized nature of
this deployment.