casaos-user-management-service-updates (#2185)
* 🎨 feat(user-management): Enhance CasaOS user management service

This commit introduces several improvements to the CasaOS User Management service:

- Clarifies the purpose and functionality of the service in the Docker Compose configuration
- Adds more detailed volume mappings and container capabilities to enable system integration and administrative tasks
- Enhances the application metadata for better integration with the CasaOS app store and installation process
- Provides additional environment variables for configuring the Flask application and default admin credentials

These changes aim to improve the overall user experience and system integration for the CasaOS User Management service.

* 🔧 feat(user-management): Map port 5000 on the host to 5000 on the container

The changes in this commit update the port mapping in the `docker-compose.yml` file
for the `big-bear-casaos-user-management` service. The previous mapping of port
7681 on the host to 7681 on the container has been changed to map port 5000 on
the host to 5000 on the container. This change is necessary to ensure the
user management service is accessible on the correct port.