✨ feat(upsnap): Add Upsnap application configuration (#2246)
Adds the configuration files for the Upsnap application, including the
config.json and docker-compose.yml files. This allows the Upsnap
application to be easily deployed and managed within the BigBearCasaOS
platform.

The Upsnap application is a simple wake on lan web app written with
SvelteKit, Go and PocketBase. This integration provides users with an
easy way to set up and manage the Upsnap application within their
BigBearCasaOS environment.