update-youtube-links (#2555)
* 🔖 feat(config): add YouTube links for Pterodactyl Panel and Wings

Adds YouTube links to the `config.json` files for both the Pterodactyl Panel
and Wings components. This will provide users with a direct link to the
project's YouTube channel, which can be useful for learning and support
purposes.

* ✨ feat(wallos): Add YouTube link to config.json

Adds a new YouTube link to the wallos app configuration. This
allows users to easily access the app's video content directly
from the app details page.

* ✨ feat(umbrel-os): Add YouTube link to config.json

Adds a YouTube link to the `config.json` file for the Umbrel OS
application. This will allow users to easily access a video
tutorial or overview of the Umbrel OS.

* ✨ feat(nexterm): add YouTube link to config

Adds a YouTube link to the `nexterm` app configuration in the
`config.json` file. This allows users to easily access a video
demonstration of the app.

* ✨ feat(coolify): Add YouTube link to config.json

The changes add a YouTube link to the `config.json` file for the Coolify
application. This allows users to access a related YouTube video from the
application.

* 🔖 feat(bookstack): add YouTube link to config

Adds a YouTube link to the BookStack configuration file. This
allows users to easily access the BookStack YouTube tutorial
from within the application.

* ✨ feat(config): add YouTube link to app config

Adds a YouTube link to the `config.json` file for the
"big-bear-casaos-user-management" app. This will allow users to
easily access a video tutorial or other relevant content about the
app.