✨ feat(uisp): Add UISP application configuration (#2616)
* ✨ feat(uisp): Add UISP application configuration

This commit adds the configuration files for the UISP (formerly UNMS) application to the BigBearCasaOS project. The changes include:

- Added the `logo.png` file for the UISP application.
- Created the `config.json` file with the necessary metadata for the UISP application, including the version, image, and links.
- Added the `docker-compose.yml` file to define the UISP service, including the container image, environment variables, volumes, and port mappings.
- Included the CasaOS-specific configuration in the `docker-compose.yml` file, such as supported architectures, main service, description, and port mapping.

These changes will allow the UISP application to be easily installed and managed within the BigBearCasaOS platform.

* 📝 docs(uisp): update documentation link in docker-compose.yml

The changes update the documentation link in the `docker-compose.yml` file for the
UISP app. The previous link was outdated, and the new link provides more
accurate and up-to-date information for users before installing the UISP app.

* 🔧 feat(uisp): Update docker-compose configuration for UISP

The changes update the docker-compose configuration file for the UISP
application. The name of the application is changed from `big-bear-unms`
to `big-bear-uisp` to accurately reflect the application being deployed.
Additionally, the service definitions have been updated to reflect the
changes in the application name.