add-big-bear-lyrionmusicserver (#2646)
* ✨ feat(lyrionmusicserver): add docker-compose and config files

This commit adds the necessary files for the LyriOn Music Server application to be
deployed using Docker Compose. The key changes include:

- Add `docker-compose.yml` file to define the service and its configuration
- Add `config.json` file to provide metadata about the application
- Add `logo.png` file as the application's icon

These changes will allow the LyriOn Music Server to be easily deployed and
integrated into the BigBearCasaOS platform, providing users with a seamless
streaming audio server experience.

* 📝 docs(lyrion-music-server): update installation instructions

Update the installation instructions for the Lyrion Music Server
to point to the correct community thread with more detailed
documentation.