🔧 feat(config): update database host to use docker service name
The changes update the database host configuration in the `config.toml` file to use the Docker service name `big-bear-libredesk-db` instead of the local host `127.0.0.1`. This ensures that the application can connect to the database running in a Docker container.

Additionally, the `LIBREDESK_SYSTEM_USER_PASSWORD` environment variable in the `docker-compose.yml` file has been updated to a more secure value.