✨ feat(libredesk): Add configuration and Docker Compose setup (#3205)
* ✨ feat(libredesk): Add configuration and Docker Compose setup

This commit introduces the configuration and Docker Compose setup for the libredesk
application. The key changes include:

- Added the `config.toml` file with configurations for the app, HTTP server, file
  upload provider, database, Redis, message processing, notifications, automation,
  auto-assigner, and conversation management.
- Added the `docker-compose.yml` file to define the services for the libredesk
  application, including the main libredesk service, PostgreSQL database, and Redis.
- Configured the volumes, environment variables, and port mappings for the
  libredesk service and its dependencies.
- Included CasaOS-specific configuration for the volumes and environment variables.

These changes provide a comprehensive setup for running the libredesk application
in a containerized environment, making it easier to deploy and manage the
application.

* ✨ feat(libredesk): add installation instructions and documentation

Adds a new section to the `docker-compose.yml` file for the `libredesk`
app, providing users with a link to important documentation and
installation instructions before they install the app.

* 🔧 feat(config): update database host to use docker service name

The changes update the database host configuration in the `config.toml` file to use the Docker service name `big-bear-libredesk-db` instead of the local host `127.0.0.1`. This ensures that the application can connect to the database running in a Docker container.

Additionally, the `LIBREDESK_SYSTEM_USER_PASSWORD` environment variable in the `docker-compose.yml` file has been updated to a more secure value.