🎨 feat(README): Add app store sponsorship opportunities (#3396)
* 🎨 feat(README): Add app store sponsorship opportunities

Adds a new section to the README file detailing the app store sponsorship
opportunities for developers and companies. This includes three tiers of
sponsorship with different benefits for each level. The goal is to provide
visibility and reach for applications in the BigBearCasaOS App Store, which
is the default option on IceWhale hardware and software platforms.

* 📝 docs(README): improve app store description and sponsorship details

The changes in this commit improve the description of the BigBearCasaOS App Store and the details around sponsorship opportunities. The key changes are:

- Clarify that the app store is pre-installed on IceWhale's hardware and software platforms, including CasaOS and ZimaOS.
- Emphasize that the app store is fully maintained by BigBearTechWorld and the BigBearCommunity as a separate entity from IceWhale, while maintaining a collaborative partnership.
- Update the description of the app store being one of the largest third-party app stores and the most up-to-date for CasaOS and ZimaOS.