✨ feat(morphos): Add Morphos application to CasaOS (#3424)
* ✨ feat(morphos): Add Morphos application to CasaOS

This commit adds the Morphos application to the CasaOS platform. Morphos is a self-hosted file converter server that provides a user-friendly web UI for converting files privately, without relying on third-party services.

The changes include:

- Adding a `docker-compose.yml` file to define the Morphos service and its configuration, including volume mappings, port mappings, and healthcheck settings.
- Adding a `config.json` file to provide metadata for the CasaOS app store and installation process, including the application ID, version, image, and links to documentation and support resources.
- Adding a `logo.png` file to provide the application icon for the CasaOS app store.

These changes allow users to easily install and use the Morphos application within the CasaOS platform, providing a secure and private file conversion solution.

* 📝 docs(morphos): update installation instructions

Update the installation instructions for Morphos in the
docker-compose.yml file. The new link provides more up-to-date
information for users.

* 🎨 feat(morphos): add screenshots and thumbnail

Adds new screenshot and thumbnail images for the Morphos app in the
docker-compose.yml file. This provides users with a better visual
representation of the app's features and functionality.