✨ feat: Upgrade dependencies and Node.js version
This commit upgrades the `glob` dependency to the latest version `11.0.2` and
updates the Node.js version used in the CI workflows from 18 to 20. These
changes are made to ensure the project is using the most up-to-date and
secure dependencies, as well as taking advantage of the latest features and
performance improvements in Node.js 20.