✨ feat: Upgrade dependencies and Node.js version (#3702)
* ✨ feat: Upgrade dependencies and Node.js version

This commit upgrades the `glob` dependency to the latest version `11.0.2` and
updates the Node.js version used in the CI workflows from 18 to 20. These
changes are made to ensure the project is using the most up-to-date and
secure dependencies, as well as taking advantage of the latest features and
performance improvements in Node.js 20.

* ✨ feat(pnpm-lock.yaml): Update dependencies

Update the dependencies in the `pnpm-lock.yaml` file to the latest versions. The key changes are:

- Update `jest-runtime` to version `29.7.0`.
- Update various `@commitlint` dependencies to version `19.8.0`.
- Update `@eslint-community/eslint-utils` to version `4.6.1`.
- Update `@eslint/config-array` to version `0.20.0`.
- Update `@eslint/core` to version `0.13.0`.
- Update `@eslint/eslintrc` to version `3.3.1`.
- Update `@types/node` to version `20.17.31`.
- Update `collect-v8-coverage` to version `1.0.2`.
- Update `istanbul-lib-coverage` to version `3.2.2`.
- Update `istanbul-lib-instrument` to version `6.0.3`.
- Update `istanbul-lib-report` to version `3.0.1`.

These updates ensure that the project is using the latest versions of the dependencies, which may include bug fixes, performance improvements, and new features.