🔧 refactor: Update developer information in app metadata (#3746)
The changes update the developer information in the metadata for the "Stalwart Mail Server" and "Poste.io" apps. The "Stalwart Mail Server" app is now attributed to "stalwartlabs" instead of "bigbeartechworld", and the "Poste.io" app is now attributed to "analogic" instead of "bigbeartechworld". This change is made to accurately reflect the actual developers of these applications.