feat(btop): Upgrade btop to version 0.1.2 (#3980)
This commit upgrades the btop application to version 0.1.2. The changes
include updating the version number in the config.json file and the
docker-compose.yml file. This upgrade is necessary to ensure the
application is running the latest stable version and to take advantage
of any bug fixes or new features that may have been introduced.