feat(docker-compose): update postgres image to latest version (#3988)
The changes update the Postgres image used in the Immich Docker Compose configuration to the latest version, which includes the VectorChord and PGVectors extensions. This will provide improved performance and functionality for the Immich application. The changes also include an optional environment variable to specify the storage type for the database, which can be useful if the database is not stored on SSDs.