Immich downgrade postgres image to 14 (#3997)
* 🐛 fix(docker-compose): downgrade postgres image to 14

The changes in this commit are to downgrade the Postgres image used in the Immich
application's Docker Compose configuration from version 16 to version 14. This
change is necessary to address compatibility issues that have been reported with
the newer version of Postgres.

* ✨ feat(deps): disable Renovate updates for Postgres image

Disables Renovate updates for the `ghcr.io/immich-app/postgres` Docker image.
This is to prevent unintended updates to the Postgres version, which could
potentially break the application.