✨ feat(wishlist): Add new wishlist app configuration (#4092)
* ✨ feat(wishlist): Add new wishlist app configuration

Adds the configuration files for the BigBearWishlist app, including the
Docker Compose file, config.json, and associated assets like screenshots
and the logo. This new app will provide a self-hosted wishlist
application that users can share with friends and family.

* 🔖 docs(wishlist): update installation instructions

Updates the installation instructions for the Wishlist app in the
docker-compose.yml file. The new instructions point to the correct
community thread for the Wishlist app.